pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=26.3.1",
    "dotenv>=0.9.9",
    "isort>=8.0.1",
    "langchain>=1.2.15",
    "langchain-openai>=1.1.12",
    "langsmith>=0.7.29",
    "python-dotenv>=1.2.2",
]
